+++ tools/origami-tools-helpers/index.js
@@ -2,5 +2,4 @@
 
 export * as files from "./files.js"
 import log from "./log.js"
-import { constructPolyfillUrl } from "./construct-polyfill-url.js";
-export {log, constructPolyfillUrl}
+export {log}